Asignación de calificiones
Creación de la vista encargada de la asignación de calificaciones para los estudiantes. La vista permite almacenar la nota de forma cuantitativa para los casos en que la evaluación sea de este tipo o sea cualitativa (con su equivalente cualitativo). Creación de permisos para usuarios, a fin de que las URL's solo puedan ser vistas por el usuario específico. Ajuste en las rutas para obviar los métodos del recurso "QualificationController" que no serán usadas. Ajustes menores en el menú principal y en los permisos de las funcionalidades.

+++ app/Http/Controllers/EvaluationController.php
@@ -107,6 +107,10 @@ class EvaluationController extends Controller
     {
         $evaluation = Evaluation::findOrFail($id);
 
+        if ( $evaluation->plan->user->first()->id != Auth::user()->id ) {
+            abort(403);
+        }
+
         $plan = Plan::findOrFail($evaluation->plan_id);
 
         $conceptual = ConceptualSection::findOrFail($plan->conceptual_section_id);
@@ -132,6 +136,11 @@ class EvaluationController extends Controller
     public function edit($id)
     {
         $evaluation = Evaluation::findOrFail($id);
+
+        if ( Carbon::now()->gt($evaluation->start_date) ||
+             $evaluation->plan->user->first()->id != Auth::user()->id ) {
+            abort(403);
+        }
 
         $plans = Auth::user()->plans;
 
@@ -189,6 +198,11 @@ class EvaluationController extends Controller
     public function destroy($id)
     {
         $evaluation = Evaluation::findOrFail($id);
+
+        if ( Carbon::now()->gt($evaluation->start_date) ||
+             $evaluation->plan->user->first()->id != Auth::user()->id ) {
+            abort(403);
+        }
 
         $evaluation->delete();
